send-sms/requirements.txt
Flask
gevent
kavenegar
PyYAML
